requirements.txt
playwright==1.49.0
pytest==8.3.4
pytest-playwright==0.6.2
pytest-asyncio==0.24.0
allure-pytest==2.13.5
pyyaml==6.0.2
python-dotenv==1.0.1
